New class (DataFileWriter) to handle writing of processed data.     Code restructuring of data processing modules and apps
