+++ src/base/OrderedEventHandler.hpp
@@ -62,6 +62,10 @@ namespace PETSYS {
 		virtual void report() {
 			this->sink->report();
 		};
+
+		virtual void resetCounters() {
+			this->sink->resetCounters();
+		};
 		
 	protected:
 		virtual EventBuffer<TEventOutput> * handleEvents(EventBuffer<TEventInput> *inBuffer) = 0;
